Fix error on "cabal test"
```
cabal: Cannot test the package bellman-ford-0.1.0.0 because none of the
components are available to build: the test suite 'bellman-ford-test' is not
available because the solver did not find a plan that included the test
suites. Force the solver to enable this for all packages by adding the line
'tests: True' to the 'cabal.project.local' file.
```